Added Project Logos, and set Z-Index for Nav
The nav was not fixed, actually...which was the primary problem for
display. Too, other elements decided to draw over it even when it was
fixed, so we reset the z-index. Also, logos from the people mockups (as
in, project logos), added.

Robert, you didn't add really any relevant information here that people
could have just read in the title.