New Photos??
See what you all think. I'm open to whatever, but I noticed that with
the scenes varying, it would look kind of awkward with just a standard
black gradient backdrop on two photos. I could fix that for one of
them...I could not fix that for me yet, because I hate every single
photo I come across (but almost replaced it with person_no.png).