Started the carousel. Everybody jump on, now!
I try to bring some joy to the masses with my work. Some. Don't get
carried away, now.

Anyway, so we've been talking about adding a carousel, and this will
take care of some of the basic additions. You'll see I also set up an
image array, so we can probably get this done fairly quickly. This isn't
actually that difficult...probably the hardest part is going to be
making it pretty. That said, we'll do anything to make this look a
little better here.