Transparent IndigoBox Logo, Join Additions
Wow, this commit is terrible. That's why I put one of the smaller things
first, because it least is something related to a logo that we know and
trust.

This covers the addition of a clipboard image and the transparent
IndigoBox logo on Snowshoes Join. I also tried to use the shade of
orange we are using as a background color, but...I don't know. This is
just...sad. That's all.