Snowshoes Join Edits Reinstated, Button Fix
So, it looks like I was using two different versions of main1.css to
modify the site...and somehow out of nowhere, the header changes and all
the work from last night disappeared on the live copy. Hopefully this
fixes it. Good news, however, is that it was using a version of
main1.css with the fix for the button, so I was able to bring that back
here.