Adjusted Source for Old Snowshoes Build Page
This page is not linked nor referenced anywhere, and should probably be
removed altogether from the project files due to its age (and terrible
writing). That said, it had an impact on recent tests due to a reference
to an image that no longer existed, and that has been repaired now.